Day 34: Improve floor and device update validation

+++ backend/src/main/java/com/preethi/smartcampus/service/BuildingService.java
@@ -1,6 +1,7 @@
 package com.preethi.smartcampus.service;
-import java.util.Optional;
+
 import com.preethi.smartcampus.entity.Building;
+import com.preethi.smartcampus.exception.ResourceNotFoundException;
 import com.preethi.smartcampus.repository.BuildingRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -9,7 +10,6 @@ import java.util.List;
 
 @Service
 public class BuildingService {
-    
 
     @Autowired
     private BuildingRepository buildingRepository;
@@ -17,18 +17,44 @@ public class BuildingService {
     public List<Building> getAllBuildings() {
         return buildingRepository.findAll();
     }
-     public Optional<Building> getBuildingById(Long id) {
-    return buildingRepository.findById(id);
-}
+
+    public Building getBuildingById(Long id) {
+
+        return buildingRepository.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException(
+                                "Building not found with id: " + id
+                        )
+                );
+    }
 
     public Building saveBuilding(Building building) {
         return buildingRepository.save(building);
     }
-    public Building updateBuilding(Long id, Building updatedBuilding) {
-    updatedBuilding.setId(id);
-    return buildingRepository.save(updatedBuilding);
-}
-public void deleteBuilding(Long id) {
-    buildingRepository.deleteById(id);
-}
+
+    public Building updateBuilding(
+            Long id,
+            Building updatedBuilding) {
+
+        if (!buildingRepository.existsById(id)) {
+            throw new ResourceNotFoundException(
+                    "Building not found with id: " + id
+            );
+        }
+
+        updatedBuilding.setId(id);
+
+        return buildingRepository.save(updatedBuilding);
+    }
+
+    public void deleteBuilding(Long id) {
+
+        if (!buildingRepository.existsById(id)) {
+            throw new ResourceNotFoundException(
+                    "Building not found with id: " + id
+            );
+        }
+
+        buildingRepository.deleteById(id);
+    }
 }
